app/web_api/requirements.txt
fastapi
uvicorn[standard]
PyJWT
passlib[bcrypt]
httpx
pydantic